+++ C/.LZ76c.d
@@ -1,0 +1,17 @@
+.LZ76c.o .LZ76c.d: LZ76c.c LZ76c.h utils.h /opt/MATLAB/R2024b/extern/include/mat.h \
+ /opt/MATLAB/R2024b/extern/include/matrix.h \
+ /opt/MATLAB/R2024b/extern/include/tmwtypes.h khashl.h mt64.h
+
+LZ76c.h:
+
+utils.h:
+
+/opt/MATLAB/R2024b/extern/include/mat.h:
+
+/opt/MATLAB/R2024b/extern/include/matrix.h:
+
+/opt/MATLAB/R2024b/extern/include/tmwtypes.h:
+
+khashl.h:
+
+mt64.h:
