+++ C/.LZ76c_benchmark.d
@@ -1,0 +1,18 @@
+.LZ76c_benchmark.o .LZ76c_benchmark.d: LZ76c_benchmark.c LZ76c.h utils.h \
+ /opt/MATLAB/R2024b/extern/include/mat.h \
+ /opt/MATLAB/R2024b/extern/include/matrix.h \
+ /opt/MATLAB/R2024b/extern/include/tmwtypes.h khashl.h mt64.h
+
+LZ76c.h:
+
+utils.h:
+
+/opt/MATLAB/R2024b/extern/include/mat.h:
+
+/opt/MATLAB/R2024b/extern/include/matrix.h:
+
+/opt/MATLAB/R2024b/extern/include/tmwtypes.h:
+
+khashl.h:
+
+mt64.h:
