+++ C/.LZ76c_randmv.d
@@ -1,0 +1,18 @@
+.LZ76c_randmv.o .LZ76c_randmv.d: LZ76c_randmv.c /opt/MATLAB/R2021a/extern/include/mat.h \
+ /opt/MATLAB/R2021a/extern/include/matrix.h \
+ /opt/MATLAB/R2021a/extern/include/tmwtypes.h mt64.h LZ76c.h utils.h \
+ khashl.h
+
+/opt/MATLAB/R2021a/extern/include/mat.h:
+
+/opt/MATLAB/R2021a/extern/include/matrix.h:
+
+/opt/MATLAB/R2021a/extern/include/tmwtypes.h:
+
+mt64.h:
+
+LZ76c.h:
+
+utils.h:
+
+khashl.h:
