+++ C/.LZ78c.d
@@ -1,0 +1,17 @@
+.LZ78c.o .LZ78c.d: LZ78c.c LZ78c.h utils.h /opt/MATLAB/R2024b/extern/include/mat.h \
+ /opt/MATLAB/R2024b/extern/include/matrix.h \
+ /opt/MATLAB/R2024b/extern/include/tmwtypes.h khashl.h mt64.h
+
+LZ78c.h:
+
+utils.h:
+
+/opt/MATLAB/R2024b/extern/include/mat.h:
+
+/opt/MATLAB/R2024b/extern/include/matrix.h:
+
+/opt/MATLAB/R2024b/extern/include/tmwtypes.h:
+
+khashl.h:
+
+mt64.h:
