+++ C/.LZ78c_benchmark.d
@@ -1,0 +1,18 @@
+.LZ78c_benchmark.o .LZ78c_benchmark.d: LZ78c_benchmark.c LZ78c.h utils.h \
+ /opt/MATLAB/R2024b/extern/include/mat.h \
+ /opt/MATLAB/R2024b/extern/include/matrix.h \
+ /opt/MATLAB/R2024b/extern/include/tmwtypes.h khashl.h mt64.h
+
+LZ78c.h:
+
+utils.h:
+
+/opt/MATLAB/R2024b/extern/include/mat.h:
+
+/opt/MATLAB/R2024b/extern/include/matrix.h:
+
+/opt/MATLAB/R2024b/extern/include/tmwtypes.h:
+
+khashl.h:
+
+mt64.h:
