+++ C/.mt64.d
@@ -1,0 +1,3 @@
+.mt64.o .mt64.d: mt64.c mt64.h
+
+mt64.h:
