--- C/LZ76c_matlab.c
@@ -1,3 +1,0 @@
-#include "LZ76c_matlab.h"
-
-#include <string.h>
