--- C/LZ76c_matlab.h
@@ -1,7 +1,0 @@
-#ifndef LZC76_MATLAB_H
-#define LZC76_MATLAB_H
-
-#include "mex.h"
-#include "LZ76c.h"
-
-#endif // LZ76C_MATLAB_H
