+++ C/LZ76c_x_mex.c
@@ -1,4 +1,4 @@
-#include "LZ76c_matlab.h"
+#include "LZ76c.h"
 
 // Main function
